+++ vk/defs/src/lib.rs
@@ -118,10 +118,10 @@ pub const VK_MAX_MEMORY_HEAPS: usize = 16;
 pub const VK_MAX_EXTENSION_NAME_SIZE: usize = 256;
 pub const VK_MAX_DESCRIPTION_SIZE: usize = 256;
 
-pub type VkPipelineCacheHeaderVersion = usize;
+pub type VkPipelineCacheHeaderVersion = i32;
 pub const VK_PIPELINE_CACHE_HEADER_VERSION_ONE: VkPipelineCacheHeaderVersion = 1;
 
-pub type VkResult = isize;
+pub type VkResult = i32;
 pub const VK_SUCCESS: VkResult = 0;
 pub const VK_NOT_READY: VkResult = 1;
 pub const VK_TIMEOUT: VkResult = 2;
@@ -150,7 +150,7 @@ pub const VK_ERROR_INVALID_SHADER_NV: VkResult = -100_0012_000;
 pub const VK_ERROR_OUT_OF_POOL_MEMORY_KHR: VkResult = -100_0069_000;
 pub const VK_ERROR_INVALID_EXTERNAL_HANDLE_KHR: VkResult = -100_0072_003;
 
-pub type VkStructureType = isize;
+pub type VkStructureType = i32;
 pub const VK_STRUCTURE_TYPE_APPLICATION_INFO: VkStructureType = 0;
 pub const VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO: VkStructureType = 1;
 pub const VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO: VkStructureType = 2;
@@ -331,17 +331,17 @@ pub const VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_ADVANCED_STATE_CREATE_INFO_EXT:
 pub const VK_STRUCTURE_TYPE_PIPELINE_COVERAGE_TO_COLOR_STATE_CREATE_INFO_NV: VkStructureType = 100_0149_000;
 pub const VK_STRUCTURE_TYPE_PIPELINE_COVERAGE_MODULATION_STATE_CREATE_INFO_NV: VkStructureType = 100_0152_000;
 
-pub type VkSystemAllocationScope = isize;
+pub type VkSystemAllocationScope = i32;
 pub const VK_SYSTEM_ALLOCATION_SCOPE_COMMAND: VkSystemAllocationScope = 0;
 pub const VK_SYSTEM_ALLOCATION_SCOPE_OBJECT: VkSystemAllocationScope = 1;
 pub const VK_SYSTEM_ALLOCATION_SCOPE_CACHE: VkSystemAllocationScope = 2;
 pub const VK_SYSTEM_ALLOCATION_SCOPE_DEVICE: VkSystemAllocationScope = 3;
 pub const VK_SYSTEM_ALLOCATION_SCOPE_INSTANCE: VkSystemAllocationScope = 4;
 
-pub type VkInternalAllocationType = isize;
+pub type VkInternalAllocationType = i32;
 pub const VK_INTERNAL_ALLOCATION_TYPE_EXECUTABLE: VkInternalAllocationType = 0;
 
-pub type VkFormat = isize;
+pub type VkFormat = i32;
 pub const VK_FORMAT_UNDEFINED: VkFormat = 0;
 pub const VK_FORMAT_R4G4_UNORM_PACK8: VkFormat = 1;
 pub const VK_FORMAT_R4G4B4A4_UNORM_PACK16: VkFormat = 2;
@@ -536,32 +536,32 @@ pub const VK_FORMAT_PVRTC1_4BPP_SRGB_BLOCK_IMG: VkFormat = 1000054005;
 pub const VK_FORMAT_PVRTC2_2BPP_SRGB_BLOCK_IMG: VkFormat = 1000054006;
 pub const VK_FORMAT_PVRTC2_4BPP_SRGB_BLOCK_IMG: VkFormat = 1000054007;
 
-pub type VkImageType = isize;
+pub type VkImageType = i32;
 pub const VK_IMAGE_TYPE_1D: VkImageType = 0;
 pub const VK_IMAGE_TYPE_2D: VkImageType = 1;
 pub const VK_IMAGE_TYPE_3D: VkImageType = 2;
 
-pub type VkImageTiling = isize;
+pub type VkImageTiling = i32;
 pub const VK_IMAGE_TILING_OPTIMAL: VkImageTiling = 0;
 pub const VK_IMAGE_TILING_LINEAR: VkImageTiling = 1;
 
-pub type VkPhysicalDeviceType = isize;
+pub type VkPhysicalDeviceType = i32;
 pub const VK_PHYSICAL_DEVICE_TYPE_OTHER: VkPhysicalDeviceType = 0;
 pub const VK_PHYSICAL_DEVICE_TYPE_INTEGRATED_GPU: VkPhysicalDeviceType = 1;
 pub const VK_PHYSICAL_DEVICE_TYPE_DISCRETE_GPU: VkPhysicalDeviceType = 2;
 pub const VK_PHYSICAL_DEVICE_TYPE_VIRTUAL_GPU: VkPhysicalDeviceType = 3;
 pub const VK_PHYSICAL_DEVICE_TYPE_CPU: VkPhysicalDeviceType = 4;
 
-pub type VkQueryType = isize;
+pub type VkQueryType = i32;
 pub const VK_QUERY_TYPE_OCCLUSION: VkQueryType = 0;
 pub const VK_QUERY_TYPE_PIPELINE_STATISTICS: VkQueryType = 1;
 pub const VK_QUERY_TYPE_TIMESTAMP: VkQueryType = 2;
 
-pub type VkSharingMode = isize;
+pub type VkSharingMode = i32;
 pub const VK_SHARING_MODE_EXCLUSIVE: VkSharingMode = 0;
 pub const VK_SHARING_MODE_CONCURRENT: VkSharingMode = 1;
 
-pub type VkImageLayout = isize;
+pub type VkImageLayout = i32;
 pub const VK_IMAGE_LAYOUT_UNDEFINED: VkImageLayout = 0;
 pub const VK_IMAGE_LAYOUT_GENERAL: VkImageLayout = 1;
 pub const VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL: VkImageLayout = 2;
@@ -574,7 +574,7 @@ pub const VK_IMAGE_LAYOUT_PREINITIALIZED: VkImageLayout = 8;
 pub const VK_IMAGE_LAYOUT_PRESENT_SRC_KHR: VkImageLayout = 100_0001_002;
 pub const VK_IMAGE_LAYOUT_SHARED_PRESENT_KHR: VkImageLayout = 100_0111_000;
 
-pub type VkImageViewType = isize;
+pub type VkImageViewType = i32;
 pub const VK_IMAGE_VIEW_TYPE_1D: VkImageViewType = 0;
 pub const VK_IMAGE_VIEW_TYPE_2D: VkImageViewType = 1;
 pub const VK_IMAGE_VIEW_TYPE_3D: VkImageViewType = 2;
@@ -583,7 +583,7 @@ pub const VK_IMAGE_VIEW_TYPE_1D_ARRAY: VkImageViewType = 4;
 pub const VK_IMAGE_VIEW_TYPE_2D_ARRAY: VkImageViewType = 5;
 pub const VK_IMAGE_VIEW_TYPE_CUBE_ARRAY: VkImageViewType = 6;
 
-pub type VkComponentSwizzle = isize;
+pub type VkComponentSwizzle = i32;
 pub const VK_COMPONENT_SWIZZLE_IDENTITY: VkComponentSwizzle = 0;
 pub const VK_COMPONENT_SWIZZLE_ZERO: VkComponentSwizzle = 1;
 pub const VK_COMPONENT_SWIZZLE_ONE: VkComponentSwizzle = 2;
@@ -592,11 +592,11 @@ pub const VK_COMPONENT_SWIZZLE_G: VkComponentSwizzle = 4;
 pub const VK_COMPONENT_SWIZZLE_B: VkComponentSwizzle = 5;
 pub const VK_COMPONENT_SWIZZLE_A: VkComponentSwizzle = 6;
 
-pub type VkVertexInputRate = isize;
+pub type VkVertexInputRate = i32;
 pub const VK_VERTEX_INPUT_RATE_VERTEX: VkVertexInputRate = 0;
 pub const VK_VERTEX_INPUT_RATE_INSTANCE: VkVertexInputRate = 1;
 
-pub type VkPrimitiveTopology = isize;
+pub type VkPrimitiveTopology = i32;
 pub const VK_PRIMITIVE_TOPOLOGY_POINT_LIST: VkPrimitiveTopology = 0;
 pub const VK_PRIMITIVE_TOPOLOGY_LINE_LIST: VkPrimitiveTopology = 1;
 pub const VK_PRIMITIVE_TOPOLOGY_LINE_STRIP: VkPrimitiveTopology = 2;
@@ -609,17 +609,17 @@ pub const VK_PRIMITIVE_TOPOLOGY_TRIANGLE_LIST_WITH_ADJACENCY: VkPrimitiveTopolog
 pub const VK_PRIMITIVE_TOPOLOGY_TRIANGLE_STRIP_WITH_ADJACENCY: VkPrimitiveTopology = 9;
 pub const VK_PRIMITIVE_TOPOLOGY_PATCH_LIST: VkPrimitiveTopology = 10;
 
-pub type VkPolygonMode = isize;
+pub type VkPolygonMode = i32;
 pub const VK_POLYGON_MODE_FILL: VkPolygonMode = 0;
 pub const VK_POLYGON_MODE_LINE: VkPolygonMode = 1;
 pub const VK_POLYGON_MODE_POINT: VkPolygonMode = 2;
 pub const VK_POLYGON_MODE_FILL_RECTANGLE_NV: VkPolygonMode = 100_0153_000;
 
-pub type VkFrontFace = isize;
+pub type VkFrontFace = i32;
 pub const VK_FRONT_FACE_COUNTER_CLOCKWISE: VkFrontFace = 0;
 pub const VK_FRONT_FACE_CLOCKWISE: VkFrontFace = 1;
 
-pub type VkCompareOp = isize;
+pub type VkCompareOp = i32;
 pub const VK_COMPARE_OP_NEVER: VkCompareOp = 0;
 pub const VK_COMPARE_OP_LESS: VkCompareOp = 1;
 pub const VK_COMPARE_OP_EQUAL: VkCompareOp = 2;
@@ -629,7 +629,7 @@ pub const VK_COMPARE_OP_NOT_EQUAL: VkCompareOp = 5;
 pub const VK_COMPARE_OP_GREATER_OR_EQUAL: VkCompareOp = 6;
 pub const VK_COMPARE_OP_ALWAYS: VkCompareOp = 7;
 
-pub type VkStencilOp = isize;
+pub type VkStencilOp = i32;
 pub const VK_STENCIL_OP_KEEP: VkStencilOp = 0;
 pub const VK_STENCIL_OP_ZERO: VkStencilOp = 1;
 pub const VK_STENCIL_OP_REPLACE: VkStencilOp = 2;
@@ -639,7 +639,7 @@ pub const VK_STENCIL_OP_INVERT: VkStencilOp = 5;
 pub const VK_STENCIL_OP_INCREMENT_AND_WRAP: VkStencilOp = 6;
 pub const VK_STENCIL_OP_DECREMENT_AND_WRAP: VkStencilOp = 7;
 
-pub type VkLogicOp = isize;
+pub type VkLogicOp = i32;
 pub const VK_LOGIC_OP_CLEAR: VkLogicOp = 0;
 pub const VK_LOGIC_OP_AND: VkLogicOp = 1;
 pub const VK_LOGIC_OP_AND_REVERSE: VkLogicOp = 2;
@@ -657,7 +657,7 @@ pub const VK_LOGIC_OP_OR_INVERTED: VkLogicOp = 13;
 pub const VK_LOGIC_OP_NAND: VkLogicOp = 14;
 pub const VK_LOGIC_OP_SET: VkLogicOp = 15;
 
-pub type VkBlendFactor = isize;
+pub type VkBlendFactor = i32;
 pub const VK_BLEND_FACTOR_ZERO: VkBlendFactor = 0;
 pub const VK_BLEND_FACTOR_ONE: VkBlendFactor = 1;
 pub const VK_BLEND_FACTOR_SRC_COLOR: VkBlendFactor = 2;
@@ -678,7 +678,7 @@ pub const VK_BLEND_FACTOR_ONE_MINUS_SRC1_COLOR: VkBlendFactor = 16;
 pub const VK_BLEND_FACTOR_SRC1_ALPHA: VkBlendFactor = 17;
 pub const VK_BLEND_FACTOR_ONE_MINUS_SRC1_ALPHA: VkBlendFactor = 18;
 
-pub type VkBlendOp = isize;
+pub type VkBlendOp = i32;
 pub const VK_BLEND_OP_ADD: VkBlendOp = 0;
 pub const VK_BLEND_OP_SUBTRACT: VkBlendOp = 1;
 pub const VK_BLEND_OP_REVERSE_SUBTRACT: VkBlendOp = 2;
@@ -731,7 +731,7 @@ pub const VK_BLEND_OP_RED_EXT: VkBlendOp = 100_0148_043;
 pub const VK_BLEND_OP_GREEN_EXT: VkBlendOp = 100_0148_044;
 pub const VK_BLEND_OP_BLUE_EXT: VkBlendOp = 100_0148_045;
 
-pub type VkDynamicState = isize;
+pub type VkDynamicState = i32;
 pub const VK_DYNAMIC_STATE_VIEWPORT: VkDynamicState = 0;
 pub const VK_DYNAMIC_STATE_SCISSOR: VkDynamicState = 1;
 pub const VK_DYNAMIC_STATE_LINE_WIDTH: VkDynamicState = 2;
@@ -744,23 +744,23 @@ pub const VK_DYNAMIC_STATE_STENCIL_REFERENCE: VkDynamicState = 8;
 pub const VK_DYNAMIC_STATE_VIEWPORT_W_SCALING_NV: VkDynamicState = 100_0087_000;
 pub const VK_DYNAMIC_STATE_DISCARD_RECTANGLE_EXT: VkDynamicState = 100_0099_000;
 
-pub type VkFilter = isize;
+pub type VkFilter = i32;
 pub const VK_FILTER_NEAREST: VkFilter = 0;
 pub const VK_FILTER_LINEAR: VkFilter = 1;
 pub const VK_FILTER_CUBIC_IMG: VkFilter = 100_0015_000;
 
-pub type VkSamplerMipmapMode = isize;
+pub type VkSamplerMipmapMode = i32;
 pub const VK_SAMPLER_MIPMAP_MODE_NEAREST: VkSamplerMipmapMode = 0;
 pub const VK_SAMPLER_MIPMAP_MODE_LINEAR: VkSamplerMipmapMode = 1;
 
-pub type VkSamplerAddressMode = isize;
+pub type VkSamplerAddressMode = i32;
 pub const VK_SAMPLER_ADDRESS_MODE_REPEAT: VkSamplerAddressMode = 0;
 pub const VK_SAMPLER_ADDRESS_MODE_MIRRORED_REPEAT: VkSamplerAddressMode = 1;
 pub const VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE: VkSamplerAddressMode = 2;
 pub const VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_BORDER: VkSamplerAddressMode = 3;
 pub const VK_SAMPLER_ADDRESS_MODE_MIRROR_CLAMP_TO_EDGE: VkSamplerAddressMode = 4;
 
-pub type VkBorderColor = isize;
+pub type VkBorderColor = i32;
 pub const VK_BORDER_COLOR_FLOAT_TRANSPARENT_BLACK: VkBorderColor = 0;
 pub const VK_BORDER_COLOR_INT_TRANSPARENT_BLACK: VkBorderColor = 1;
 pub const VK_BORDER_COLOR_FLOAT_OPAQUE_BLACK: VkBorderColor = 2;
@@ -768,7 +768,7 @@ pub const VK_BORDER_COLOR_INT_OPAQUE_BLACK: VkBorderColor = 3;
 pub const VK_BORDER_COLOR_FLOAT_OPAQUE_WHITE: VkBorderColor = 4;
 pub const VK_BORDER_COLOR_INT_OPAQUE_WHITE: VkBorderColor = 5;
 
-pub type VkDescriptorType = isize;
+pub type VkDescriptorType = i32;
 pub const VK_DESCRIPTOR_TYPE_SAMPLER: VkDescriptorType = 0;
 pub const VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER: VkDescriptorType = 1;
 pub const VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE: VkDescriptorType = 2;
@@ -781,32 +781,32 @@ pub const VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC: VkDescriptorType = 8;
 pub const VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC: VkDescriptorType = 9;
 pub const VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT: VkDescriptorType = 10;
 
-pub type VkAttachmentLoadOp = isize;
+pub type VkAttachmentLoadOp = i32;
 pub const VK_ATTACHMENT_LOAD_OP_LOAD: VkAttachmentLoadOp = 0;
 pub const VK_ATTACHMENT_LOAD_OP_CLEAR: VkAttachmentLoadOp = 1;
 pub const VK_ATTACHMENT_LOAD_OP_DONT_CARE: VkAttachmentLoadOp = 2;
 
-pub type VkAttachmentStoreOp = isize;
+pub type VkAttachmentStoreOp = i32;
 pub const VK_ATTACHMENT_STORE_OP_STORE: VkAttachmentStoreOp = 0;
 pub const VK_ATTACHMENT_STORE_OP_DONT_CARE: VkAttachmentStoreOp = 1;
 
-pub type VkPipelineBindPoint = isize;
+pub type VkPipelineBindPoint = i32;
 pub const VK_PIPELINE_BIND_POINT_GRAPHICS: VkPipelineBindPoint = 0;
 pub const VK_PIPELINE_BIND_POINT_COMPUTE: VkPipelineBindPoint = 1;
 
-pub type VkCommandBufferLevel = isize;
+pub type VkCommandBufferLevel = i32;
 pub const VK_COMMAND_BUFFER_LEVEL_PRIMARY: VkCommandBufferLevel = 0;
 pub const VK_COMMAND_BUFFER_LEVEL_SECONDARY: VkCommandBufferLevel = 1;
 
-pub type VkIndexType = isize;
+pub type VkIndexType = i32;
 pub const VK_INDEX_TYPE_UINT16: VkIndexType = 0;
 pub const VK_INDEX_TYPE_UINT32: VkIndexType = 1;
 
-pub type VkSubpassContents = isize;
+pub type VkSubpassContents = i32;
 pub const VK_SUBPASS_CONTENTS_INLINE: VkSubpassContents = 0;
 pub const VK_SUBPASS_CONTENTS_SECONDARY_COMMAND_BUFFERS: VkSubpassContents = 1;
 
-pub type VkObjectType = isize;
+pub type VkObjectType = i32;
 pub const VK_OBJECT_TYPE_UNKNOWN: VkObjectType = 0;
 pub const VK_OBJECT_TYPE_INSTANCE: VkObjectType = 1;
 pub const VK_OBJECT_TYPE_PHYSICAL_DEVICE: VkObjectType = 2;
